data.aceadvisors.org/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
distlib==0.3.6
Django==4.2.1
django-environ==0.10.0
django-filter==23.2
django-smart-selects==1.6.0
djangorestframework==3.14.0
env==0.1.0
filelock==3.12.0
mysql-connector-python==8.0.33
mysqlclient==2.1.1
pdf2image==1.16.3
Pillow==9.5.0
pipenv==2023.5.19
platformdirs==3.5.1
protobuf==3.20.3
python-magic==0.4.27
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.23.0
virtualenv-clone==0.5.7
Wand==0.6.11
whitenoise==6.4.0
